abstracted items, times, values
